requirements.txt
pandas
tqdm
python-dotenv
requests
